Get things working after the mass-rename
