+++ rails-autoscale-core/lib/rails_autoscale/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module RailsAutoscale
-  VERSION = "1.0.0.rc1"
+  VERSION = "0.0.1"
 end
