+++ rails-autoscale-delayed_job/lib/rails_autoscale/delayed_job/version.rb
@@ -2,6 +2,6 @@
 
 module RailsAutoscale
   module DelayedJob
-    VERSION = "1.0.0.rc1"
+    VERSION = "0.0.1"
   end
 end
