+++ rails-autoscale-que/lib/rails_autoscale/que/version.rb
@@ -2,6 +2,6 @@
 
 module RailsAutoscale
   module Que
-    VERSION = "1.0.0.rc1"
+    VERSION = "0.0.1"
   end
 end
