+++ rails-autoscale-resque/lib/rails_autoscale/resque/version.rb
@@ -2,6 +2,6 @@
 
 module RailsAutoscale
   module Resque
-    VERSION = "1.0.0.rc1"
+    VERSION = "0.0.1"
   end
 end
