+++ rails-autoscale-sidekiq/lib/rails_autoscale/sidekiq/version.rb
@@ -2,6 +2,6 @@
 
 module RailsAutoscale
   module Sidekiq
-    VERSION = "1.0.0.rc1"
+    VERSION = "0.0.1"
   end
 end
