+++ rails-autoscale-web/lib/rails-autoscale-web.rb
@@ -1,3 +1,3 @@
 # frozen_string_literal: true
 
-require "rails_autoscale/rails"
+require "rails_autoscale/web"
