+++ rails-autoscale-web/lib/rails_autoscale/web/version.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module RailsAutoscale
-  module Rails
-    VERSION = "1.0.0.rc1"
+  module Web
+    VERSION = "0.0.1"
   end
 end
